requirements.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
blessings==1.7
bqplot==0.12.39
branca==0.6.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colour==0.1.5
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
decouple==0.0.7
distlib==0.3.6
Django==4.0.6
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.13.1
djangorestframework-gis==1.0
earthengine-api==0.1.357
ee==0.2
eerepr==0.0.4
executing==1.2.0
filelock==3.12.0
folium==0.14.0
fonttools==4.40.0
future==0.18.3
geemap==0.22.1
geocoder==1.38.1
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.23.2
ipyleaflet==0.17.3
ipython
ipytree==0.2.2
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.5.19
platformdirs==3.5.1
plotly==5.15.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
psycopg2-binary
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
pyperclip==1.8.2
pyshp==2.3.1
python-box==7.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.1
pyzmq==25.1.0
ratelim==0.1.6
requests==2.31.0
rsa==4.9
scooby==0.7.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
tenacity==8.2.2
tornado==6.3.2
traitlets==5.9.0
traittypes==0.2.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
widgetsnbextension==4.0.7
xyzservices==2023.5.0
